(+) moved serviceAvroFormatter to standalone file
